requirement.txt
llama-cpp-python>=0.2.23
# For Raspberry Pi GPIO:
RPi.GPIO>=0.7.0